+++ semcore/data-table/__tests__/data-table-header.browser-test.tsx
@@ -384,6 +384,34 @@ test.describe('One level header - Sorting', () => {
     expect(afterSecondSortWidths[3]).toEqual(initialWidths[3]);
   });
 
+  test('Verify sorting  not activates interactive with interactive element in same cell', async ({ page }) => {
+    const standPath = 'stories/components/data-table/tests/examples/header-tests/sorting-with-interactive.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+    await page.setContent(htmlContent);
+
+    const popper = page.locator('[data-ui-name="DescriptionTooltip.Popper"]');
+    const trigger = page.locator('[data-ui-name="DescriptionTooltip.Trigger"]');
+    const column = page.locator('[data-ui-name="Head.Column"]');
+    await page.keyboard.press('Tab');
+    await page.keyboard.press('Enter');
+
+    await expect(popper).not.toBeVisible();
+    await page.keyboard.press('Enter');
+    await expect(column.first()).not.toHaveAttribute('aria-sort');
+    await expect(popper).toBeVisible();
+    await expect(popper).toBeFocused();
+    await page.keyboard.press('Escape');
+    await expect(trigger).toBeFocused();
+
+    await page.keyboard.press('Escape');
+    await page.keyboard.press('ArrowRight');
+
+    await column.first().hover();
+    await trigger.click();
+    await expect(popper).toBeVisible();
+    await expect(column.first()).not.toHaveAttribute('aria-sort');
+  });
+
   test('Verify mouse sorting without changing size', async ({ page }) => {
     const standPath = 'stories/components/data-table/docs/examples/sorting.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');