+++ stories/components/data-table/tests/examples/header-tests/table-with-1tf-and diff-elements.tsx
@@ -114,6 +114,7 @@ const Demo = () => {
                 title='Go to our awesome article'
                 data-test-id='interactive-icon'
                 color='icon-secondary-neutral'
+                onClick={(e: any) => e.stopPropagation()}
               />
             </>
           ),